Perbaikan notifikasi

+++ client/src/components/Main/layouts/ButtonDelete.vue
@@ -8,15 +8,31 @@ export default {
   props : ['data_id', 'content'],
   methods : {
     deleteData(id){
-      this.$axios({
-        headers : {
-          token : localStorage.token
-        },
-        method : 'delete',
-        url : `/${this.content}/admin/${id}`
+      this.$swal.fire({
+        title: 'Are you sure?',
+        text: "You won't be able to revert this!",
+        icon: 'warning',
+        showCancelButton: true,
+        confirmButtonColor: '#3085d6',
+        cancelButtonColor: '#d33',
+        confirmButtonText: 'Yes, delete it!'
+      }).then((result) => {
+          if (result.value) {
+          return this.$axios({
+            headers : {
+              token : localStorage.token
+            },
+            method : 'delete',
+            url : `/${this.content}/admin/${id}`
+          })
+          }
       })
       .then(response => {
         console.log(response)
+        this.$toast.fire({
+          icon: 'success',
+          title: 'Product has been deleted'
+        })
         if(this.content == 'users'){
           this.$store.dispatch('getAdmin')
         }else{
