+++ client/src/components/Main/layouts/ButtonLogout.vue
@@ -9,8 +9,23 @@ export default {
   name : 'ButtonLogout',
   methods : {
     logout(){
-      localStorage.clear()
-      this.$router.push({ name : 'Login' })
+      this.$swal.fire({
+        title: 'Are you sure?',
+        icon: 'warning',
+        showCancelButton: true,
+        confirmButtonColor: '#3085d6',
+        cancelButtonColor: '#d33',
+        confirmButtonText: 'Yes, logout!'
+      }).then((result) => {
+        if (result.value) {
+          this.$toast.fire({
+            icon: 'success',
+            title: 'Successfully Logout!'
+          })
+        }
+        localStorage.clear()
+        this.$router.push({ name : 'Login' })
+      })
     }
   }
 }
